+++ src/gui/tools/include/brush.h
@@ -4,8 +4,8 @@
 #include <tool.h>
 
 struct Brush: public Tool {
-    void onPress(Layout& layout, const Vector2i& pos) override;
-    void onMove(Layout& layout, const Vector2i& oldPos, const Vector2i& newPos) override;
+    void onPress(Layout& layout, const Vector2f& pos) override;
+    void onMove(Layout& layout, const Vector2f& oldPos, const Vector2f& newPos) override;
 };
 
 #endif // BRUSH
