+++ src/gui/tools/src/tool.cc
@@ -5,11 +5,11 @@
 Tool::~Tool()
 {}
 
-void Tool::onMove(Layout& layout, const Vector2i& oldPos, const Vector2i& newPos)
+void Tool::onMove(Layout& layout, const Vector2f& oldPos, const Vector2f& newPos)
 {}
 
-void Tool::onPress(Layout& layout, const Vector2i& pos)
+void Tool::onPress(Layout& layout, const Vector2f& pos)
 {}
 
-void Tool::onRelease(Layout& layout, const Vector2i& pos)
+void Tool::onRelease(Layout& layout, const Vector2f& pos)
 {}
